do not compress pulp data -- RPMs are already compressed
this gives a quite huge time improvement:

    [root@sat61]# time tar cavf pulp.tar /var/lib/pulp/
    real	4m33.872s
    user	0m7.305s
    sys	0m42.243s

    [root@sat61]# time tar cavf pulp.tar.gz /var/lib/pulp/
    real	18m51.409s
    user	15m53.221s
    sys	0m57.508s

also, while at it, don't be verbose when calling tar

Getting millions of files listed while backing up and restoring is
pretty annoying.  If needed, this should be converted to a --verbose
option to katello-{backup,restore}, but let's just drop it for now.

Does only give a minimal perfomance boost in my test VM:

    [root@sat61]# time tar caf pulp.tar /var/lib/pulp/
    real	4m28.705s
    user	0m7.421s
    sys	0m35.732s

    [root@sat61]# time tar cavf pulp.tar /var/lib/pulp/
    real	4m33.872s
    user	0m7.305s
    sys	0m42.243s

    [root@sat61]# time tar caf pulp.tar.gz /var/lib/pulp/
    real	18m44.553s
    user	16m2.624s
    sys	0m50.475s

    [root@sat61]# time tar cavf pulp.tar.gz /var/lib/pulp/
    real	18m51.409s
    user	15m53.221s
    sys	0m57.508s

But makes the output way nicer to read.

https://bugzilla.redhat.com/show_bug.cgi?id=1283578